+++ cypress/e2e/showLotto.cy.js
@@ -1,43 +1,6 @@
-import { ALERT_MESSAGE } from '../../src/js/constants';
-
-describe('로또 테스트', () => {
+describe('로또 랜더링 테스트', () => {
   beforeEach(() => {
     cy.visit('../../index.html');
-  });
-  describe('로또 구입 금액의 단위는 1000원이다.', () => {
-    const alertInput = () => {
-      cy.on('window:alert', (text) => {
-        expect(text).to.contains(ALERT_MESSAGE.NOT_ONE_THOUSAND_UNIT);
-      });
-    };
-    it('10000 입력한 상태에서 확인 버튼을 클릭한다.', () => {
-      cy.clickPurchaseBtn(10000);
-    });
-    it('2000 입력한 상태에서 enter를 누른다.', () => {
-      cy.pressEnter(2000);
-    });
-    it('1234 입력한 상태에서 확인 버튼을 클릭했을 때 alert 창이 뜬다.', () => {
-      cy.clickPurchaseBtn(1234);
-      alertInput();
-    });
-    it('2500 입력한 상태에서 enter를 눌렀을 때 alert 창이 뜬다.', () => {
-      cy.pressEnter(2500);
-      alertInput();
-    });
-  });
-  describe('로또 구입 금액은 1000원 이상이다.', () => {
-    it('0원 입력하면 alert창이 뜬다.', () => {
-      cy.clickPurchaseBtn(0);
-      cy.on('window:alert', (text) => {
-        expect(text).to.contains(ALERT_MESSAGE.NOT_OVER_ONE_THOUSAND);
-      });
-    });
-    it('700원 입력하면 alert창이 뜬다.', () => {
-      cy.clickPurchaseBtn(700);
-      cy.on('window:alert', (text) => {
-        expect(text).to.contains(ALERT_MESSAGE.NOT_OVER_ONE_THOUSAND);
-      });
-    });
   });
   it('구입하지 않은 상태에서 구입 금액 칸과 확인 버튼만 보인다.', () => {
     cy.getByDataCy('lotto-purchase-input').should('exist');
@@ -74,15 +37,17 @@ describe('로또 테스트', () => {
     });
   });
   describe('구입한 상태에서 번호 보기 토글 버튼을 클릭했을 때 중복되지 않은 랜덤한 6개의 숫자를 화면에 표시한다.', () => {
+    const showLottoItemNumbers = () =>
+      cy.getByDataCy('lotto-item-numbers').should('have.css', 'display', 'block');
     it('5000원을 구매한 상태에서 번호 보기 토글 버튼을 눌렀을 때 중복되지 않은 랜덤한 6개의 숫자가 화면에 표시된다.', () => {
       cy.clickPurchaseBtn(5000);
       cy.get('.switch').click();
-      cy.getByDataCy('lotto-item-numbers').should('have.css', 'display', 'block');
+      showLottoItemNumbers();
     });
     it('15000원을 구매한 상태에서 번호 보기 토글 버튼을 눌렀을 때 중복되지 않은 랜덤한 6개의 숫자가 화면에 표시된다.', () => {
       cy.clickPurchaseBtn(15000);
       cy.get('.switch').click();
-      cy.getByDataCy('lotto-item-numbers').should('have.css', 'display', 'block');
+      showLottoItemNumbers();
     });
   });
 });
